+++ mitra_activitypub/src/adapters/users.rs
@@ -1,13 +1,32 @@
 use mitra_config::Config;
 use mitra_models::{
     accounts::types::User,
+    activitypub::queries::save_actor,
     database::{DatabaseClient, DatabaseError},
     profiles::queries::delete_profile,
 };
+use mitra_services::media::MediaServer;
 
 use crate::{
+    actors::builders::build_local_actor,
+    authority::Authority,
     builders::delete_person::prepare_delete_person,
 };
+
+// NOTE: not called when emojis are updated
+pub async fn create_or_update_local_actor(
+    config: &Config,
+    db_client: &mut impl DatabaseClient,
+    account: &User,
+) -> Result<(), DatabaseError> {
+    let authority = Authority::from(&config.instance());
+    let media_server = MediaServer::new(config);
+    let actor = build_local_actor(&authority, &media_server, account)?;
+    let actor_json = serde_json::to_value(&actor)
+        .expect("actor should be serializable");
+    save_actor(db_client, &actor.id, &actor_json, account.id).await?;
+    Ok(())
+}
 
 pub async fn delete_user(
     config: &Config,